Consolidate logging utilities to a new `dwave.cloud.utils.logging` module
